refactor: pagedto location
